+++ ICSharpCode.CodeConverter/Util/StringExtensions.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Diagnostics;
+using System.Text.RegularExpressions;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
@@ -456,5 +457,16 @@ namespace ICSharpCode.CodeConverter.Util
         {
             return originalContainingReplacement.Substring(0, originalContainingReplacement.Length - replacement.Key.Length) + replacement.Value;
         }
+    }
+
+    /// <summary>
+    /// Copied from https://github.com/code-cracker/code-cracker/blob/330f7ad217d6aae17a07a0675f52bcaa9a03d956/src/Common/CodeCracker.Common/Extensions/GeneratedCodeAnalysisExtensions.cs
+    /// Apache-2.0 license
+    /// </summary>
+    public static class GeneratedCodeAnalysisExtensions
+    {
+        public static bool IsGeneratedFile(this string filePath) =>
+            Regex.IsMatch(filePath, @"(\\service|\\TemporaryGeneratedFile_.*|\\assemblyinfo|\\assemblyattributes|\.(g\.i|g|designer|generated|assemblyattributes))\.(cs|vb)$",
+                RegexOptions.IgnoreCase);
     }
 }